+++ R/bcFuns.R
@@ -32,14 +32,14 @@
 #' EBPH \tab 21 \tab Exponential Body Proportional Hypothesis \cr
 #' ESPH \tab 22 \tab Exponential Scale Proportional Hypothesis \cr
 #' }
-#' @return A function that can be used to predict length at previous age (Li) given length-at-capture (Lc), hard-part radius-at-age i (Ri), and hard-part radius-at-capture (Rc).  In addition, some functions/models may require the previous age (agei) and the age-at-capture (agec), certain parameters related to the biological intercept (R0p & L0p), or certain parameters estimated from various regression models (a,b,c,A,B,C).  See source for more information.
+#' @return A function that can be used to predict length at previous age (Li) given length-at-capture (Lcap), hard-part radius-at-age i (Ri), and hard-part radius-at-capture (Rcap). In addition, some functions/models may require the previous age (agei) and the age-at-capture (agec), certain parameters related to the biological intercept (R0p & L0p), or certain parameters estimated from various regression models (a,b,c,A,B,C). See source for more information.
 #'
 #' @author Derek H. Ogle, \email{derek@@derekogle.com}
 #'
 #' @section IFAR Supplement: \url{http://derekogle.com/IFAR/supplements/backcalculation/}
 #'
 #' @references
-#' Vigliola, L. and M.G. Meekan.  2009.  The back-calculation of fish growth from otoliths.  pp. 174-211.  in B.S. Green et al. (editors).  Tropical Fish Otoliths: Information for Assessment, Management and Ecology.  Review: Methods and Technologies in Fish Biology and Fisheries 11.  Springer.  [Was (is?) available from https://www.researchgate.net/publication/226394736_The_Back-Calculation_of_Fish_Growth_From_Otoliths.]
+#' Vigliola, L. and M.G. Meekan. 2009. The back-calculation of fish growth from otoliths. pp. 174-211. in B.S. Green et al. (editors). Tropical Fish Otoliths: Information for Assessment, Management and Ecology. Review: Methods and Technologies in Fish Biology and Fisheries 11. Springer. [Was (is?) available from https://www.researchgate.net/publication/226394736_The_Back-Calculation_of_Fish_Growth_From_Otoliths.]
 #'
 #' @keywords manip
 #'
@@ -60,12 +60,11 @@
 #' @export
 bcFuns <- function(BCM,verbose=FALSE) {
   ## Do some checking
-  if (missing(BCM)) stop("A back-calculation function must be chosen with 'BCM'",call.=FALSE)
-  if (length(BCM)>1) stop("Only one value may be given to 'BCM'",
-                          call.=FALSE)
+  if (missing(BCM)) STOP("A back-calculation function must be chosen with 'BCM'")
+  if (length(BCM)>1) STOP("Only one value may be given to 'BCM'")
   if (is.numeric(BCM)) {
     ## Function declared numerically
-    if (BCM<1 | BCM>22) stop("BCM number must be between 1 and 22 inclusive.",call.=FALSE)
+    if (BCM<1 | BCM>22) STOP("BCM number must be between 1 and 22 inclusive.")
   } else {
     ## Function declared by name
     BCM.nms <- c("DALE","FRALE","BI","LBI","BPH","LBPH","TVG","SPH",
@@ -80,7 +79,7 @@ bcFuns <- function(BCM,verbose=FALSE) {
       msg <- paste(strwrap(paste("'BCM' must be one of:",
                                  paste(BCM.nms,collapse=", ")),
                            width=62),collapse="\n")
-      stop(msg,call.=FALSE)
+      STOP(msg)
     } else {
       # All is good ... convert string to numeric
       BCM <- BCM.nums[which(BCM.nms %in% BCM)]
@@ -90,69 +89,69 @@ bcFuns <- function(BCM,verbose=FALSE) {
   ## identify the functions
   if (BCM==1) {
       if (verbose) message("You chose the BCM1 or DALE model.\n\n")
-      function(Lc,Ri,Rc) { (Ri/Rc)*Lc }
+      function(Lcap,Ri,Rcap) { (Ri/Rcap)*Lcap }
   } else if (BCM==2) {
       if (verbose) message("You chose the BCM2 or FRALE model.\n\n")
-      function(Lc,Ri,Rc,a) { a+(Lc-a)*(Ri/Rc) }
+      function(Lcap,Ri,Rcap,a) { a+(Lcap-a)*(Ri/Rcap) }
   } else if (BCM==3) {
       if (verbose) message("You chose the BCM3, BI, or LBI model.\n\n")
-      function(Lc,Ri,Rc,L0p,R0p) { Lc+(Ri-Rc)*(Lc-L0p)/(Rc-R0p) }
+      function(Lcap,Ri,Rcap,L0p,R0p) { Lcap+(Ri-Rcap)*(Lcap-L0p)/(Rcap-R0p) }
   } else if (BCM==4) {
       if (verbose) message("You chose the BCM4 or LBPH model.\n\n")
-      function(Lc,Ri,Rc,a,b) { (a+b*Ri)*Lc/(a+b*Rc) }
+      function(Lcap,Ri,Rcap,a,b) { (a+b*Ri)*Lcap/(a+b*Rcap) }
   } else if (BCM==5) {
-      stop("The BCM5 (TVG) function is not yet implemented.",call.=FALSE)
+      STOP("The BCM5 (TVG) function is not yet implemented.",call.=FALSE)
   } else if (BCM==6) {
       if (verbose) message("You chose the BCM6 or LSPH model.\n\n")
-      function(Lc,Ri,Rc,A,B) { (Ri/Rc*(A+B*Lc)-A)/B }
+      function(Lcap,Ri,Rcap,A,B) { (Ri/Rcap*(A+B*Lcap)-A)/B }
   } else if (BCM==7) {
       if (verbose) message("You chose the BCM7, AE, or AESPH model.\n\n")
-      function(Lc,Ri,Rc,agei,agec,a,b,c) { -a/b+(Lc+a/b+c/b*agec)*Ri/Rc-c/b*agei }
+      function(Lcap,Ri,Rcap,agei,agec,a,b,c) { -a/b+(Lcap+a/b+c/b*agec)*Ri/Rcap-c/b*agei }
   } else if (BCM==8) {
       if (verbose) message("You chose the BCM8 or AEBPH model.\n\n")
-      function(Lc,Ri,Rc,agei,agec,A,B,C) { (A+B*Ri+C*agei)/(A+B*Rc+C*agec)*Lc }
+      function(Lcap,Ri,Rcap,agei,agec,A,B,C) { (A+B*Ri+C*agei)/(A+B*Rcap+C*agec)*Lcap }
   } else if (BCM==9) {
       if (verbose) message("You chose the BCM9 or MONA model.\n\n")
-      function(Lc,Ri,Rc,c) { Lc*((Ri/Rc)^c) }
+      function(Lcap,Ri,Rcap,c) { Lcap*((Ri/Rcap)^c) }
   } else if (BCM==10) {
       if (verbose) message("You chose the BCM10 or MONA-BPH model.\n\n")
       # Same as BCM9 but uses nls results to estimate c
-      function(Lc,Ri,Rc,c) { Lc*((Ri/Rc)^c) }
+      function(Lcap,Ri,Rcap,c) { Lcap*((Ri/Rcap)^c) }
   } else if (BCM==11) {
       if (verbose) message("You chose the BCM11 or MONA-SPH model.\n\n")
-      function(Lc,Ri,Rc,C) { Lc*((Ri/Rc)^(1/C)) }
+      function(Lcap,Ri,Rcap,C) { Lcap*((Ri/Rcap)^(1/C)) }
   } else if (BCM==12) {
       if (verbose) message("You chose the BCM12 or WAKU model.\n\n")
-      function(Lc,Ri,Rc,L0p,R0) {
-        exp(log(L0p) + ((log(Lc)-log(L0p))*(log(Ri)-log(R0)))/(log(Rc)-log(R0)))
+      function(Lcap,Ri,Rcap,L0p,R0) {
+        exp(log(L0p) + ((log(Lcap)-log(L0p))*(log(Ri)-log(R0)))/(log(Rcap)-log(R0)))
       }
   } else if (BCM==13) {
       if (verbose) message("You chose the BCM13 or FRY model.\n\n")
-      function(Lc,Ri,Rc,L0,R0,a) {
-        a + exp(log(L0-a) + ((log(Lc-a)-log(L0-a))*(log(Ri)-log(R0)))/(log(Rc)-log(R0))) 
+      function(Lcap,Ri,Rcap,L0,R0,a) {
+        a + exp(log(L0-a) + ((log(Lcap-a)-log(L0-a))*(log(Ri)-log(R0)))/(log(Rcap)-log(R0))) 
       }
   } else if (BCM==14) {
       if (verbose) message("You chose the BCM14, MF, or ABI model.\n\n")
-      function(Lc,Ri,Rc,L0p,R0p,a) {
-        a + exp(log(L0p-a) + ((log(Lc-a)-log(L0p-a))*(log(Ri)-log(R0p)))/(log(Rc)-log(R0p)))
+      function(Lcap,Ri,Rcap,L0p,R0p,a) {
+        a + exp(log(L0p-a) + ((log(Lcap-a)-log(L0p-a))*(log(Ri)-log(R0p)))/(log(Rcap)-log(R0p)))
       }
   } else if (BCM==15) {
       if (verbose) message("You chose the BCM15, FRY-BPH, or ABPH model.\n\n")
-      function(Lc,Ri,Rc,a,b,c) { (a+b*Ri^c)/(a+b*Rc^c)*Lc }
+      function(Lcap,Ri,Rcap,a,b,c) { (a+b*Ri^c)/(a+b*Rcap^c)*Lcap }
   } else if (BCM==16) {
       if (verbose) message("You chose the BCM16, FRY-SPH, or ASPH model.\n\n")
-      function(Lc,Ri,Rc,a,b,c) { a+(Lc-a)*((Ri/Rc)^c) }
+      function(Lcap,Ri,Rcap,a,b,c) { a+(Lcap-a)*((Ri/Rcap)^c) }
   } else if (BCM==17) {
       if (verbose) message("You chose the BCM17 or QBPH model.\n\n")
-      function(Lc,Ri,Rc,a,b,c) { (a+b*Ri+c*(Ri^2))/(a+b*Rc+c*(Rc^2))*Lc }
+      function(Lcap,Ri,Rcap,a,b,c) { (a+b*Ri+c*(Ri^2))/(a+b*Rcap+c*(Rcap^2))*Lcap }
   } else if (BCM==18) {
       if (verbose) message("You chose the BCM18 or QSPH model.\n\n")
-      function(Lc,Ri,Rc,a,b,c) { 
+      function(Lcap,Ri,Rcap,a,b,c) { 
         cf1 <- b
         cf2 <- c
-        Li <- numeric(length(Lc))
+        Li <- numeric(length(Lcap))
         for (i in seq_along(Li)) {
-          cf0 <- a-((Ri[i]/Rc[i])*(a+b*Lc[i]+c*Lc[i]^2))
+          cf0 <- a-((Ri[i]/Rcap[i])*(a+b*Lcap[i]+c*Lcap[i]^2))
           roots <- Re(polyroot(c(cf0,cf1,cf2)))
           Li[i] <- roots[which(sign(roots)==1)]
         }
@@ -160,14 +159,14 @@ bcFuns <- function(BCM,verbose=FALSE) {
       }
   } else if (BCM==19) {
       if (verbose) message("You chose the BCM19 or PBPH model.\n\n")
-      function(Lc,Ri,Rc,a) {
+      function(Lcap,Ri,Rcap,a) {
         # a must be a vector of coefficients from polynomial regression
         exps <- 0:(length(a)-1)
-        Li <- numeric(length(Lc))
-        for (i in seq_along(Lc)) {
+        Li <- numeric(length(Lcap))
+        for (i in seq_along(Lcap)) {
           num <- sum( a*Ri[i]^exps )
-          denom <- sum( a*Rc[i]^exps )
-          Li[i] <- num/denom*Lc[i]
+          denom <- sum( a*Rcap[i]^exps )
+          Li[i] <- num/denom*Lcap[i]
         }
         Li
       }
@@ -175,20 +174,20 @@ bcFuns <- function(BCM,verbose=FALSE) {
       # Note that this is a function that should be used when finding
       #   a root, not to actually back-calculate
       if (verbose) message("You chose the BCM20 or PSPH model.\n\n")
-      function(Lc,Ri,Rc,a) {
+      function(Lcap,Ri,Rcap,a) {
         # a must be a vector of coefficients from polynomial regression
         exps <- 0:(length(a)-1)
-        Li <- numeric(length(Lc))
+        Li <- numeric(length(Lcap))
         for (i in seq_along(Li)) {
-          if (Ri[i]==Rc[i]) { Li[i] <- Lc[i] }
+          if (Ri[i]==Rcap[i]) { Li[i] <- Lcap[i] }
           else {
             cf <- a
-            cf[1] <- cf[1] - Ri[i]/Rc[i]*sum(a*Lc[i]^exps)
+            cf[1] <- cf[1] - Ri[i]/Rcap[i]*sum(a*Lcap[i]^exps)
             roots <- Re(polyroot(cf))
             # find only positive roots
             roots <- roots[which(sign(roots)==1)]
             # only find root less than lencap
-            roots <- roots[which(roots<=Lc[i])]
+            roots <- roots[which(roots<=Lcap[i])]
             ifelse(length(roots)!=1,Li[i] <- NA,Li[i] <- roots)
           }
         }
@@ -196,9 +195,120 @@ bcFuns <- function(BCM,verbose=FALSE) {
       }      
   } else if (BCM==21) {
       if (verbose) message("You chose the BCM21 or EBPH model.\n\n")
-      function(Lc,Ri,Rc,a,b) { exp(a+b*Ri)/exp(a+b*Rc)*Lc }
+      function(Lcap,Ri,Rcap,a,b) { exp(a+b*Ri)/exp(a+b*Rcap)*Lcap }
   } else if (BCM==22) {
       if (verbose) message("You chose the BCM22 or ESPH model.\n\n")
-      function(Lc,Ri,Rc,a) { exp(a+(log(Lc)-a)*Ri/Rc) }
+      function(Lcap,Ri,Rcap,a) { exp(a+(log(Lcap)-a)*Ri/Rcap) }
   } 
+}
+
+
+
+#' @name BCfunctions
+#' @rdname BCfunctions
+#' @title Specific back-calculation functions
+#' @description Provides a function to compute back-calculated lengths at previous ages for several of the most common back-calculation models: Dahl-Lea, Fraser-Lee, Body Proportional Hypothesis (BPH), Scale Proportional Hypothesis (SPH), Biological Intercept, and Modified Fry. See \code{\link{bcFuns}} for other back-calculation functions.
+#' @param Lcap A numeric vector of lengths-at-capture.
+#' @param Ri A numeric vector of radial measurements to the ith annulus.
+#' @param Rcap A numeric vector of total structure radii-at-capture.
+#' @param A The intercept of the radius- on length-at-capture regression.
+#' @param B The slope of the radius- on length-at-capture regression.
+#' @param a The length of the fish at the time of structure formation, the intercept of the length- on radius-at-capture regression, or, when using scales, from published “standards” for a species (Carlander 1982).
+#' @param b The slope of the length- on radius-at-capture regression.
+#' @param L0p Fish length at the \dQuote{Biological Intercept}.
+#' @param R0p Radius at the \dQuote{Biological Intercept}.
+#' @references See references in \code{\link{bcFuns}}.
+#' @seealso \code{\link{bcFuns}}.
+#' @return A back-calculated length-at-age.
+NULL
+
+#' @rdname BCfunctions
+#' @export
+DahlLea <- function(Lcap,Ri,Rcap) {
+  if (!is.numeric(Lcap)) STOP("'Lcap' must be numeric.")
+  if (any(Lcap<=0)) STOP("All 'Lcap' must be positive.")
+  if (!is.numeric(Ri)) STOP("'Ri' must be numeric.")
+  if (any(Ri<=0)) STOP("All 'Ri' must be positive.")
+  if (!is.numeric(Rcap)) STOP("'Rcap' must be numeric.")
+  if (any(Rcap<=0)) STOP("All 'Rcap' must be positive.")
+  (Ri/Rcap)*Lcap
+}
+
+#' @rdname BCfunctions
+#' @export
+FraserLee <- function(Lcap,Ri,Rcap,a) {
+  if (!is.numeric(Lcap)) STOP("'Lcap' must be numeric.")
+  if (any(Lcap<=0)) STOP("All 'Lcap' must be positive.")
+  if (!is.numeric(Ri)) STOP("'Ri' must be numeric.")
+  if (any(Ri<=0)) STOP("All 'Ri' must be positive.")
+  if (!is.numeric(Rcap)) STOP("'Rcap' must be numeric.")
+  if (any(Rcap<=0)) STOP("All 'Rcap' must be positive.")
+  if (!is.numeric(a)) STOP("'a' must be numeric.")
+  if (length(a)>1) STOP("'a' must be a single value.")
+  (Ri/Rcap)*(Lcap-a)+a
+}
+
+#' @rdname BCfunctions
+#' @export
+SPH <- function(Lcap,Ri,Rcap,A,B) {
+  if (!is.numeric(Lcap)) STOP("'Lcap' must be numeric.")
+  if (any(Lcap<=0)) STOP("All 'Lcap' must be positive.")
+  if (!is.numeric(Ri)) STOP("'Ri' must be numeric.")
+  if (any(Ri<=0)) STOP("All 'Ri' must be positive.")
+  if (!is.numeric(Rcap)) STOP("'Rcap' must be numeric.")
+  if (any(Rcap<=0)) STOP("All 'Rcap' must be positive.")
+  if (!is.numeric(A)) STOP("'A' must be numeric.")
+  if (length(A)>1) STOP("'A' must be a single value.")
+  if (!is.numeric(B)) STOP("'B' must be numeric.")
+  if (length(B)>1) STOP("'B' must be a single value.")
+  (Ri/Rcap)*(Lcap+A/B)-A/B
+}
+
+#' @rdname BCfunctions
+#' @export
+BPH <- function(Lcap,Ri,Rcap,a,b) {
+  if (!is.numeric(Lcap)) STOP("'Lcap' must be numeric.")
+  if (any(Lcap<=0)) STOP("All 'Lcap' must be positive.")
+  if (!is.numeric(Ri)) STOP("'Ri' must be numeric.")
+  if (any(Ri<=0)) STOP("All 'Ri' must be positive.")
+  if (!is.numeric(Rcap)) STOP("'Rcap' must be numeric.")
+  if (any(Rcap<=0)) STOP("All 'Rcap' must be positive.")
+  if (!is.numeric(a)) STOP("'a' must be numeric.")
+  if (length(a)>1) STOP("'a' must be a single value.")
+  if (!is.numeric(b)) STOP("'b' must be numeric.")
+  if (length(b)>1) STOP("'b' must be a single value.")
+  Lcap*((a+b*Ri)/(a+b*Rcap))
+}
+
+#' @rdname BCfunctions
+#' @export
+BiolInt <- function(Lcap,Ri,Rcap,L0p,R0p) {
+  if (!is.numeric(Lcap)) STOP("'Lcap' must be numeric.")
+  if (any(Lcap<=0)) STOP("All 'Lcap' must be positive.")
+  if (!is.numeric(Ri)) STOP("'Ri' must be numeric.")
+  if (any(Ri<=0)) STOP("All 'Ri' must be positive.")
+  if (!is.numeric(Rcap)) STOP("'Rcap' must be numeric.")
+  if (any(Rcap<=0)) STOP("All 'Rcap' must be positive.")
+  if (!is.numeric(L0p)) STOP("'L0p' must be numeric.")
+  if (length(L0p)>1) STOP("'L0p' must be a single value.")
+  if (!is.numeric(R0p)) STOP("'R0p' must be numeric.")
+  if (length(R0p)>1) STOP("'R0p' must be a single value.")
+  Lcap+(Ri-Rcap)*(Lcap-L0p)/(Rcap-R0p)
+}
+#' @rdname BCfunctions
+#' @export
+ModFry <- function(Lcap,Ri,Rcap,L0p,R0p,a) {
+  if (!is.numeric(Lcap)) STOP("'Lcap' must be numeric.")
+  if (any(Lcap<=0)) STOP("All 'Lcap' must be positive.")
+  if (!is.numeric(Ri)) STOP("'Ri' must be numeric.")
+  if (any(Ri<=0)) STOP("All 'Ri' must be positive.")
+  if (!is.numeric(Rcap)) STOP("'Rcap' must be numeric.")
+  if (any(Rcap<=0)) STOP("All 'Rcap' must be positive.")
+  if (!is.numeric(L0p)) STOP("'L0p' must be numeric.")
+  if (length(L0p)>1) STOP("'L0p' must be a single value.")
+  if (!is.numeric(R0p)) STOP("'R0p' must be numeric.")
+  if (length(R0p)>1) STOP("'R0p' must be a single value.")
+  if (!is.numeric(a)) STOP("'a' must be numeric.")
+  if (length(a)>1) STOP("'a' must be a single value.")
+  a + exp(log(L0p-a) + ((log(Lcap-a)-log(L0p-a))*(log(Ri)-log(R0p)))/(log(Rcap)-log(R0p)))
 }